# Conflicts:
#   Engine/RenderCore/OpenGL/AnimationController.cpp
#   Game/TestArea/TestPlayer.cpp

+++ Engine/RenderCore/OpenGL/Mesh.cpp
@@ -330,8 +330,8 @@ MeshDataHolder::~MeshDataHolder()
 
 void MeshDataHolder::CreateState()
 {
-	VertexCount = temp_vertices.size();
-	IndexCount = temp_indices.size();
+	VertexCount = (uint32)temp_vertices.size();
+	IndexCount = (uint32)temp_indices.size();
 
 	glGenVertexArrays(1, &VAO);
